+++ src/aigora/curriculum_graph/application/parsing/graph_parser.py
