+++ src/aigora/curriculum_graph/application/parsing/parser_errors.py
